+++ source/TestCommon/documents/release-notes/release-notes.md
@@ -6,6 +6,7 @@
   - Specify which functions to load using `FunctionAppHostSettings.Functions`.
   - Set environment variables for the process using `FunctionAppHostSettings.ProcessEnvironmentVariables`.
   - Write host process id to output when started, so we know which process to attach debugger to if needed.
+- Updated default values for `FunctionAppHostSettings` to reduce the necessary settings in  `functionapphost.settings.json`.
 - Implemented `FunctionAppHostManagerExtensions`:
   - `CheckIfFunctionWasExecuted`
   - `TriggerFunctionAsync`